+++ cw-orch-daemon/Cargo.toml
@@ -21,7 +21,7 @@ default = []
 # enable node-backed tests (ensure Docker is running)
 # run with `cargo test --jobs 1 --features node-tests`
 node-tests = []
-
+eth = ["dep:ethers-signers", "dep:ethers-core"]
 [dependencies]
 # Default deps
 cw-orch-contract-derive = { path = "../packages/cw-orch-contract-derive", version = "0.13.3" }
@@ -61,8 +61,8 @@ derive_builder = { version = "0.12.0" }
 ring = { version = "0.16.20" }
 
 # Injective dependencies
-ethers-signers = { version = "2.0.7" }
-ethers-core = { version = "2.0.7" }
+ethers-signers = { version = "2.0.7", optional = true }
+ethers-core = { version = "2.0.7", optional = true }
 
 [dev-dependencies]
 cw-orch-daemon = { path = "." }
