+++ cw-orch-daemon/src/sync/core.rs
@@ -2,7 +2,7 @@ use std::{fmt::Debug, time::Duration};
 
 use super::super::{sender::Wallet, DaemonAsync};
 use crate::{
-    queriers::{CosmWasm, DaemonQuerier, Node},
+    queriers::{DaemonQuerier, Node},
     CosmTxResponse, DaemonBuilder, DaemonError, RcDaemonState,
 };
 
@@ -10,7 +10,7 @@ use cosmrs::tendermint::Time;
 use cosmwasm_std::{Addr, Coin};
 use cw_orch_environment::{
     contract::{interface_traits::Uploadable, WasmPath},
-    environment::{ChainState, IndexResponse, TxHandler},
+    environment::{ChainState, TxHandler},
 };
 use serde::{de::DeserializeOwned, Serialize};
 use tokio::runtime::Handle;
@@ -96,33 +96,7 @@ impl TxHandler for Daemon {
     }
 
     fn upload(&self, uploadable: &impl Uploadable) -> Result<Self::Response, DaemonError> {
-        let sender = &self.daemon.sender;
-        let wasm_path = uploadable.wasm();
-
-        log::debug!("Uploading file at {:?}", wasm_path);
-
-        let file_contents = std::fs::read(wasm_path.path())?;
-        let store_msg = cosmrs::cosmwasm::MsgStoreCode {
-            sender: sender.pub_addr()?,
-            wasm_byte_code: file_contents,
-            instantiate_permission: None,
-        };
-        let result = self
-            .rt_handle
-            .block_on(sender.commit_tx(vec![store_msg], None))?;
-
-        log::info!("Uploaded: {:?}", result.txhash);
-
-        let code_id = result.uploaded_code_id().unwrap();
-
-        // wait for the node to return the contract information for this upload
-        let wasm = CosmWasm::new(self.channel());
-        while self.rt_handle.block_on(wasm.code(code_id)).is_err() {
-            self.rt_handle
-                .block_on(tokio::time::sleep(Duration::from_secs(6)));
-        }
-
-        Ok(result)
+        self.rt_handle.block_on(self.daemon.upload(uploadable))
     }
 
     fn execute<E: Serialize>(
